Handle key press/release
Since Emacs doesn't have natively these two events, we simulate the key release
with a timer. If a key get pressed in the meantime the previous timer gets
canceled, a key release of the previous pressed key will be emitted and the a
key press will be emitted for the current key pressed.