Better keypress/keyrelease handling with multiple pressed keys
- With multiple contemporary pressed keys
- Without global variables (`chip8--current-key`)
- Without global timer (`chip8--current-key-timer`)

Store the time when the key is pressed in the key's vector and set the timer,
when the timer is off check if the time passed since the key was pressed, if the
amount of time is greater or equal to the timeout then release the key.

This will support a sequence of keypress events on the same key, each of which
will override the last time the key was pressed and therefore will disable the
previously set release time.